server/pyproject.toml
[project]
name = "TermiPics"
version = "0.1.0"
description = "Server side for TermiPics, an image serving service."
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "google-auth>=2.39.0",
    "pillow>=11.2.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "supabase>=2.15.0",
]

[dependency-groups]
dev = ["ruff>=0.11.7"]

[tool.ruff]
exclude = [".git", ".venv"]
line-length = 100
indent-width = 4

[tool.ruff.lint]
# Rule categories:
# - F: Pyflakes
# - COM: flake8-commas
# - T20: flake8-print
# - LOG: flake8-logging
# - G: flake8-logging-format
# - ARG - flake8-unused-arguments
# - I: isort
# - UP: pyupgrade
# - E: pycodestyle errors
# - W: pycodestyle warnings
# - FAST: Fast API-related rules
select = ["F", "COM", "LOG", "G", "ARG", "T20", "I", "UP", "E", "W", "FAST"]

# Ignored rules:
# - UP007: Use `X | Y` in type annotations
# - COM812: Trailing comma missing (formatter handled it)
# - E501: Line too long (formatter handled it)
ignore = ["UP007", "COM812", "E501", "E"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
